Add files via upload
When the navigation frame is Earth-Centered-Earth Fixed (ECEF) frame, one might desire to project the errors on to NED (north-easth-down) frame. ECEF frame is not intuitive enough to make clear conclusions. But NED frame is more easy for performance analysis. This volume includes a few additional utilities to project the errors on to NED frame.